WIP: move ctap2 implementation from transport, fix examples
