+++ webauthn-authenticator-rs/examples/authenticate/main.rs
@@ -2,35 +2,31 @@
 extern crate tracing;
 
 use std::io::{stdin, stdout, Write};
+use std::ops::Deref;
 
-use base64urlsafedata::Base64UrlSafeData;
+use futures::executor::block_on;
+use webauthn_authenticator_rs::ctap2::CtapAuthenticator;
 use webauthn_authenticator_rs::prelude::Url;
 use webauthn_authenticator_rs::softtoken::SoftToken;
-use webauthn_authenticator_rs::transport::ctap21pre::Ctap21PreAuthenticator;
 use webauthn_authenticator_rs::transport::*;
 use webauthn_authenticator_rs::ui::{Cli, UiCallback};
 use webauthn_authenticator_rs::AuthenticatorBackend;
 use webauthn_rs_core::proto::RequestAuthenticationExtensions;
 use webauthn_rs_core::WebauthnCore as Webauthn;
 
-fn access_card<T: Token, U: UiCallback>(card: T, ui: U) -> Ctap21PreAuthenticator<T, U> {
-    info!("Card detected ...");
-
-    card.auth(ui).expect("couldn't open card")
-}
-
-fn select_transport() -> Box<dyn AuthenticatorBackend> {
-    // TODO
-    let ui = Cli {};
-
+fn select_transport<'a, U: UiCallback>(ui: &'a U) -> impl AuthenticatorBackend + 'a {
     let mut reader = AnyTransport::default();
     info!("Using reader: {:?}", reader);
 
     match reader.tokens() {
         Ok(mut tokens) => {
-            while let Some(mut card) = tokens.pop() {
-                card.init().expect("couldn't init card");
-                return Box::new(access_card(card, ui));
+            while let Some(card) = tokens.pop() {
+                let auth = block_on(CtapAuthenticator::new(card, ui));
+
+                match auth {
+                    Some(auth) => return auth,
+                    None => (),
+                }
             }
         }
         Err(e) => panic!("Error: {:?}", e),
@@ -39,19 +35,19 @@ fn select_transport() -> Box<dyn AuthenticatorBackend> {
     panic!("no card");
 }
 
-fn select_provider() -> Box<dyn AuthenticatorBackend> {
-    let mut providers: Vec<(&str, fn() -> Box<dyn AuthenticatorBackend>)> = Vec::new();
+fn select_provider<'a>(ui: &'a Cli) -> Box<dyn AuthenticatorBackend + 'a> {
+    let mut providers: Vec<(&str, fn(&'a Cli) -> Box<dyn AuthenticatorBackend>)> = Vec::new();
 
-    providers.push(("SoftToken", || Box::new(SoftToken::new().unwrap().0)));
-    providers.push(("CTAP", select_transport));
+    providers.push(("SoftToken", |_| Box::new(SoftToken::new().unwrap().0)));
+    providers.push(("CTAP", |ui| Box::new(select_transport(ui))));
 
     #[cfg(feature = "u2fhid")]
-    providers.push(("Mozilla", || {
+    providers.push(("Mozilla", |_| {
         Box::new(webauthn_authenticator_rs::u2fhid::U2FHid::default())
     }));
 
     #[cfg(feature = "win10")]
-    providers.push(("Windows 10", || {
+    providers.push(("Windows 10", |_| {
         Box::new(webauthn_authenticator_rs::win10::Win10::default())
     }));
 
@@ -77,7 +73,7 @@ fn select_provider() -> Box<dyn AuthenticatorBackend> {
                 } else {
                     let p = providers.remove((v as usize) - 1);
                     println!("Using {}...", p.0);
-                    return p.1();
+                    return p.1(&ui);
                 }
             }
             Err(_) => println!("Input was not a number"),
@@ -88,8 +84,10 @@ fn select_provider() -> Box<dyn AuthenticatorBackend> {
 
 fn main() {
     tracing_subscriber::fmt::init();
+    // TODO
+    let ui = Cli {};
 
-    let mut u = select_provider();
+    let mut u = select_provider(&ui);
 
     // WARNING: don't use this as an example of how to use the library!
     let wan = Webauthn::new_unsafe_experts_only(
