+++ webauthn-authenticator-rs/examples/nfc_token_info/core.rs
@@ -1,19 +1,17 @@
+use futures::executor::block_on;
 // use webauthn_authenticator_rs::nfc::*;
-use webauthn_authenticator_rs::{transport::*, ui::Cli};
+use webauthn_authenticator_rs::{transport::*, ui::Cli, ctap2::CtapAuthenticator};
 
 fn access_card<T: Token>(card: T) {
     info!("Card detected ...");
 
-    match card.select_any(Cli {}) {
-        Ok(Selected::FIDO_2_1_PRE(mut token)) => {
-            info!("Using token {:?}", token);
+    let auth = block_on(CtapAuthenticator::new(card, &Cli{}));
 
-            token.hack_make_cred().unwrap();
-            token.deselect_applet().unwrap();
+    match auth {
+        Some(x) => {
+            info!("Using token: {:?}", x);
         }
-        _ => {
-            unimplemented!();
-        }
+        None => unimplemented!(),
     }
 }
 
@@ -24,8 +22,7 @@ pub(crate) fn event_loop() {
 
     match reader.tokens() {
         Ok(mut tokens) => {
-            while let Some(mut card) = tokens.pop() {
-                card.init().expect("couldn't init card");
+            while let Some(card) = tokens.pop() {
                 access_card(card);
             }
         }
