+++ webauthn-authenticator-rs/src/ctap2/commands/get_assertion.rs
