+++ webauthn-authenticator-rs/src/ctap2/commands/reset.rs
