+++ webauthn-authenticator-rs/src/nfc/mod.rs
@@ -215,91 +215,6 @@ impl<'b> Transport<'b> for NFCReader {
 
         Ok(r)
     }
-
-    /*
-    async fn select_one_token<'a, U: UiCallback>(
-        &mut self,
-        ui: &'a U,
-    ) -> Result<Ctap21PreAuthenticator<'a, Self::Token, U>, WebauthnCError> {
-        // Select the first token from the list - there is no UI
-        let tokens = self.tokens()?.drain(..).filter_map(|c| {
-            async {
-                let info = c.transmit(GetInfoRequest {}, ui).await.ok()?;
-
-                Some((c, info))
-            }
-        });
-
-        match tokens.first() {
-            Some(t) => Ok(t),
-            None => Err(WebauthnCError::NoSelectedToken),
-        }
-
-        // todo!()
-        // let mut reader_states: MutexedReaderStates =
-        //     MutexedReaderStates::new(self.get_reader_states());
-        // loop {
-        //     reader_states.get_status_change(&self.ctx)?;
-
-        //     // Check every reader ...
-        //     let all_tokens: Vec<NFCCard> = reader_states
-        //         .iter()
-        //         .filter(|s| s.event_state().contains(State::PRESENT))
-        //         .filter_map(|s| NFCCard::new(&self.ctx, s).ok())
-        //         .filter_map(|mut c| {
-        //             block_on(c.init()).ok()?;
-        //             Some(c)
-        //         })
-        //         .collect();
-
-        //     let mut tasks: FuturesUnordered<_> = all_tokens
-        //         .drain(..)
-        //         .map(|mut token| {
-        //             async move {
-        //                 let info = match token.transmit(GetInfoRequest {}, ui).await {
-        //                     Ok(info) => {
-        //                         if !(info.versions.contains("FIDO_2_1_PRE")
-        //                             || info.versions.contains("FIDO_2_0")
-        //                             || info.versions.contains("FIDO_2_1"))
-        //                         {
-        //                             trace!("dropping unsupported token");
-        //                             return None;
-        //                         }
-        //                         info
-        //                     }
-        //                     Err(_) => return None,
-        //                 };
-
-        //                 // No need to selectionRequest
-        //                 Some((info, Mutex::new(token)))
-        //             }
-        //             .fuse()
-        //         })
-        //         .collect();
-
-        //     loop {
-        //         select! {
-        //             res = tasks.select_next_some() => {
-        //                 if let Some((info, mutex)) = res {
-        //                     trace!(?info);
-        //                     match mutex.into_inner() {
-        //                         Ok(guard) => return Ok(Ctap21PreAuthenticator::new(info, guard, ui)),
-        //                         _ => (),
-        //                     }
-        //                 }
-        //             }
-        //             complete => {
-        //                 // No tokens available
-        //                 return Err(WebauthnCError::NoSelectedToken);
-        //             }
-        //         }
-        //     }
-        // } // end loop.
-
-        // // Nothing picked
-        // Err(WebauthnCError::NoSelectedToken)
-    }
-    */
 }
 
 /// Transmits a single ISO 7816-4 APDU to the card.
