+++ webauthn-authenticator-rs/src/usb/responses.rs
@@ -196,6 +196,7 @@ impl TryFrom<&U2FHIDFrame> for Response {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::ctap2::commands::GetInfoResponse;
 
     #[test]
     fn init() {
